+++ erland-game-tileadventure/trunk/src/erland/game/tileadventure/GameObjectMapActionInterface.java
@@ -19,7 +19,7 @@ package erland.game.tileadventure;
  */
 
 public interface GameObjectMapActionInterface {
-    boolean isFree(GameObject obj,int x, int y, int z);
+    boolean isFree(GameObject obj,float x, float y, float z);
     Action startActionOnObject(GameObject obj, Action action);
     void endActionOnObject(GameObject obj, Action action);
     Action isActionPossibleOnObject(GameObject obj, Action action);
