+++ erland-game-tileadventure/trunk/src/erland/game/tileadventure/MapObjectEditorLevelManager.java
@@ -92,7 +92,7 @@ public class MapObjectEditorLevelManager extends LevelManager {
             obj.setContainer(getContainer());
             obj.read(objectParameters);
             obj.setImage(getPaletteImage(),obj.getBlockType()-1);
-            map.setObject(obj,obj.getPosX(),obj.getPosY(),obj.getPosZ());
+            map.setObject(obj,(int)obj.getPosX(),(int)obj.getPosY(),(int)obj.getPosZ());
             return new LevelInfo(map,info);
         }else {
             return null;
